add get_charge tool
