+++ chemicaltoolbox/rdkit/rdkit_get_charge.xml
@@ -1,0 +1,81 @@
+<tool id="ctb_rdkit_get_charge" name="Formal molecular charge" version="@TOOL_VERSION@+galaxy@GALAXY_VERSION@">
+    <description>calculated with RDKit</description>
+    <macros>
+        <token name="@TOOL_VERSION@">2020.03.4</token>
+        <token name="@GALAXY_VERSION@">0</token>
+    </macros>
+    <requirements>
+        <requirement type="package" version="@TOOL_VERSION@">rdkit</requirement>
+    </requirements>
+    <command detect_errors="exit_code">
+<![CDATA[
+        python '$__tool_directory__/get_charge.py'
+            -i '${infile}'
+            -f '${infile.ext}'
+            -o '${outfile}'
+]]>
+    </command>
+    <inputs>
+        <param name="infile" format="smi,sdf,mol2,inchi,pdb" type="data" label="Molecule data"
+            help="In SDF, SMILES, MOL2, InChI or PDB format"/>
+    </inputs>
+    <outputs>
+        <data format="tabular" name="outfile" />
+    </outputs>
+    <tests>
+        <test>
+            <param name="infile" ftype='sdf' value="sucos_cluster.sdf" />
+            <output name="outfile" ftype='tabular' file="sucos_cluster.tab" />
+        </test>
+        <test>
+            <param name="infile" ftype='pdb' value="mol.pdb" />
+            <output name="outfile" ftype='tabular' file="mol_pdb_charges.tab" />
+        </test>
+    </tests>
+    <help>
+<![CDATA[
+
+.. class:: infomark
+
+**What this tool does**
+Calculates the formal charge for a molecule and prints it to a file. If the input file contains multiple
+molecules (e.g. a SDF or SMILES file), the charge for each molecule will be written to a different line,
+in the same order as the input file.
+
+
+-----
+
+.. class:: infomark
+
+**Input**
+
+| - `SDF Format`_
+| - `SMILES Format`_
+| - `Corina MOL2`_
+| - PDB
+| - InChI format
+
+.. _SDF Format: http://en.wikipedia.org/wiki/Chemical_table_file
+.. _SMILES Format: http://en.wikipedia.org/wiki/Simplified_molecular_input_line_entry_specification
+.. _Corina MOL2: http://www.molecular-networks.com/products/corina
+
+-----
+
+.. class:: infomark
+
+ **Output**
+
+Tabular file with a single column.
+
+
+]]>
+    </help>
+    <citations>
+        <citation type="bibtex">
+            @article{rdkit,
+                author = {Greg Landrum and others},
+                title = {RDKit: Open-source cheminformatics},
+                url ={http://www.rdkit.org}
+            }</citation>
+    </citations>
+</tool>
